requirements.txt
Python 3.8+
TensorFlow 2.x
Flask
Pillow
NumPy
Matplotlib